+++ example/example/lib/stories/zwap_select_story.dart
@@ -1,3 +1,6 @@
+import 'dart:math';
+
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:zwap_design_system/atoms/atoms.dart';
 import 'package:zwap_design_system/atoms/select/zwapSelect.dart';
@@ -11,36 +14,54 @@ class ZwapSelectStory extends StatefulWidget {
 
 class _ZwapSelectStoryState extends State<ZwapSelectStory> {
   String? _selected;
-  double _top = 0;
+
+  double __top = 0;
+  set _top(double value) {
+    __top = min(max(0, value), MediaQuery.of(context).size.height);
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        _DraggableIndicator(
-          onChangeDx: (top) => setState(() => _top = top),
-        ),
-        AnimatedPositioned(
-          duration: const Duration(milliseconds: 30),
-          top: _top,
-          left: 400,
-          child: Container(
-            width: 300,
-            child: ZwapSelect(
-              canSearch: true,
-              values: Map.fromEntries(List.generate(50, (i) => MapEntry<String, String>(i.toString(), '$i-$i•$i'))),
-              fetchMoreData: (String newQuery, int pageNumber) async => {},
-              hintText: "Seleziona un elemento",
-              label: "Zwap Select",
-              callBackFunction: (value, _) => setState(() => _selected = value),
-              selected: _selected,
-              initialPageNumber: 2,
-              betweenFetchDuration: const Duration(seconds: 2),
-              onEmptyResponseDuration: const Duration(seconds: 10),
-            ),
+    return Listener(
+      onPointerSignal: (signal) {
+        if (signal is PointerScrollEvent) {
+          setState(() => _top = __top - signal.scrollDelta.dy);
+          print(signal.device);
+        }
+      },
+      child: SingleChildScrollView(
+        physics: NeverScrollableScrollPhysics(),
+        child: Container(
+          height: MediaQuery.of(context).size.height,
+          child: Stack(
+            children: [
+              _DraggableIndicator(
+                onChangeDx: (top) => setState(() => _top = top),
+              ),
+              AnimatedPositioned(
+                duration: const Duration(milliseconds: 30),
+                top: __top,
+                left: 400,
+                child: Container(
+                  width: 300,
+                  child: ZwapSelect(
+                    canSearch: true,
+                    values: Map.fromEntries(List.generate(50, (i) => MapEntry<String, String>(i.toString(), '$i-$i•$i'))),
+                    fetchMoreData: (String newQuery, int pageNumber) async => {},
+                    hintText: "Seleziona un elemento",
+                    label: "Zwap Select",
+                    callBackFunction: (value, _) => setState(() => _selected = value),
+                    selected: _selected,
+                    initialPageNumber: 2,
+                    betweenFetchDuration: const Duration(seconds: 2),
+                    onEmptyResponseDuration: const Duration(seconds: 10),
+                  ),
+                ),
+              ),
+            ],
           ),
         ),
-      ],
+      ),
     );
   }
 }