+++ example/example/lib/story_book_widget.dart
@@ -1,5 +1,6 @@
 import 'package:example/main.dart';
 import 'package:example/stories.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:zwap_design_system/atoms/atoms.dart';
 import 'package:provider/provider.dart';
@@ -26,9 +27,7 @@ class StoryBookWidget extends StatelessWidget {
                 children: stories.map((e) => _getStorySelector(e, () => context.read<StoryProvider>().currentStory = e.story)).toList(),
               ),
             ),
-            Expanded(
-              child: Builder(builder: stories.firstWhere((s) => s.story == _currentStory).builder),
-            ),
+            Expanded(child: Builder(builder: stories.firstWhere((s) => s.story == _currentStory).builder)),
           ],
         ),
       ),