+++ lib/utils/maximum_fractional_digits.dart
@@ -1,0 +1,3 @@
+extension DoubleExtension on double {
+  double getMximumFractionsDigits(int fractionalDigits) => double.parse(this.toStringAsFixed(2));
+}